+++ src/renderer_sub/swapchain_sub_renderer.cpp
@@ -154,8 +154,8 @@ namespace nugiEngine {
 		VkViewport viewport{};
 		viewport.x = 0.0f;
 		viewport.y = 0.0f;
-		viewport.width = static_cast<uint32_t>(this->width);
-		viewport.height = static_cast<uint32_t>(this->height);
+		viewport.width = static_cast<float>(this->width);
+		viewport.height = static_cast<float>(this->height);
 		viewport.minDepth = 0.0f;
 		viewport.maxDepth = 1.0f;
